layout/generic/test/test_bug240933.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=240933
-->

<head>
  <title>Test for Bug 240933</title>
  <script type="application/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>

<body>
  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=240933">
    Mozilla Bug 240933
  </a>
  <p id="display"></p>
  <div id="content" style="display: none">
  </div>

  <pre id="test">
    <script type="application/javascript">

      /** Test for Bug 240933 **/

      SimpleTest.waitForExplicitFinish();

      SimpleTest.waitForFocus(function() {
        var t = document.getElementById("t");
        synthesizeMouse(t, t.clientWidth / 2, 5, {}, window);
        is(t.selectionStart, 3, "The selection should be set before the newline");
        is(t.selectionEnd,   3, "The selection should be set before the newline");
        SimpleTest.finish();
      });

    </script>
  </pre>

  <textarea id="t" rows="10" cols="10">abc
</textarea>

</body>
</html>
